[Update]UIの調整 [Update]敵を倒したときだけタイムが増えるようにした [Update]タイムの増え方の調整 [Update]敵の配置を増やした

+++ ES Game Library/MyClass/TimeManager/TimeManager.h
@@ -26,7 +26,7 @@ public:
 	//c‚èŠÔ‚Ìæ“¾
 	inline float GetTimeLeft()  const { return limitTime - time; }
 	bool StartFlag();
-	void AddTime(int count) { limitTime += count; }
+	void AddTime(float count) { limitTime += count; }
 private:
 	inline float GetLimitTime() const { return limitTime; }
 	int GetTimeMinutes();
